+++ src/lib/components/Field/FieldLabel.js
@@ -1,19 +1,37 @@
 import React from 'react';
 import T from 'prop-types';
 
-const FieldLabel = ({ label, name, required }) => {
+import Icon from '../Icon';
+import Tooltip from '../Tooltip';
+
+const FieldLabel = (props) => {
+  const {
+    label,
+    name,
+    required,
+    tooltip,
+  } = props;
+
   if (!label) {
     return null;
   }
 
+  const tooltipId = `${name}-tooltip`;
+
   return (
     <label data-test={'label'} className={'control-label'} htmlFor={name}>
-      {
-        required
-        && <span data-test={'required'} className={'required-label'} aria-hidden>* </span>
-      }
+      {required && (
+        <span data-test={'required'} className={'required-label'} aria-hidden>* </span>
+      )}
 
       {label}
+
+      {tooltip && (
+        <span className={'field__label-tooltip'} data-for={tooltipId} data-tip>
+          <Icon name={'question-fill'} />
+          <Tooltip id={tooltipId} text={tooltip} />
+        </span>
+      )}
     </label>
   );
 };
@@ -22,11 +40,16 @@ FieldLabel.propTypes = {
   label: T.string,
   name: T.string.isRequired,
   required: T.bool,
+  tooltip: T.oneOfType([
+    T.string,
+    T.node,
+  ]),
 };
 
 FieldLabel.defaultProps = {
   label: null,
   required: false,
+  tooltip: null,
 };
 
 export default FieldLabel;