+++ src/lib/components/Field/withFieldWrapper.d.ts
@@ -2,6 +2,18 @@ import * as React from 'react';
 import * as ReduxForm from 'redux-form';
 
 export interface WrappedFieldProps {
+  /** Style class name to attach to button */
+  className?: string;
+
+  /** Disable the input */
+  disabled?: boolean;
+
+  /** Help text to render under input */
+  helpText?: string;
+
+  /** Redux form props, such as onChange and value */
+  input: ReduxForm.WrappedFieldProps;
+
   /** Label to display above input */
   label?: string;
 
@@ -11,20 +23,11 @@ export interface WrappedFieldProps {
   /** Name of field */
   name: string;
 
-  /** Help text to render under input */
-  helpText?: string;
-
-  /** Style class name to attach to button */
-  className?: string;
-
-  /** Content to render */
-  children: React.ReactNode;
-
   /** The field should use the default required validation */
   required?: boolean;
 
-  /** Disable the input */
-  disabled?: boolean;
+  /** Tooltip to display */
+  tooltip?: string | React.ReactNode;
 }
 
 declare class FieldWrapper extends React.Component<WrappedFieldProps, {}> {}